requirements.txt
alembic==1.0.5
anyjson==0.3.3
appdirs==1.4.3
APScheduler==3.5.3
arrow==0.13.0
asn1crypto==0.24.0
beautifulsoup4==4.6.3
Django==2.2.6
django-cors-headers==3.2.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djoser==2.0.3
dnspython==1.16.0
docutils==0.14
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
forex-python==0.3.3
h5py==2.8.0
Jinja2==2.10
jsonschema==2.6.0
Keras==2.3.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==15.1.0
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.3
numpy==1.15.4
numpydoc==0.8.0
oauth2==1.9.0.post1
oauth2client==3.0.0
oauthlib==3.0.0
pandas==0.23.4
pandocfilters==1.4.2
pathlib2==2.3.2
pep8==1.7.1
pickleshare==0.7.5
py==1.7.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.4.2
pyOpenSSL==18.0.0
pyparsing==2.2.2
pytest==3.8.2
pytz==2018.5
PyWavelets==1.0.1
pywin32==223
redis==2.10.6
reportlab==3.5.12
requests==2.11.1
requests-oauthlib==0.8.0
requests-toolbelt==0.7.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
scrypt==0.8.6
seaborn==0.9.0
simplejson==3.16.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.1.15
SQLAlchemy-Continuum==1.3.6
SQLAlchemy-Utils==0.32.21
sympy==1.3
tables==3.4.4
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-inet-pton==1.0.1
win-unicode-console==0.5
whitenoise==5.0.1
psycopg2==2.7.2
dj-database-url==0.5.0
gunicorn==19.7.1